pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]
